Session 3

+++ Coursework Project/KGALLA204 GP 1Coursework/KGALLA204 GP 1Coursework/cBkGround.h
@@ -13,6 +13,6 @@ class cBkGround : public cSprite
 {
 public:
 	void render();		// Default render function
-	void update(float deltaTime);		// Update method
+	void update(double deltaTime);		// Update method
 };
 #endif
